EscopoProjetoCCR/pyproject.toml
[tool.poetry]
name = "escopoprojetoccr"
version = "0.1.0"
description = ""
authors = ["edummap4412 <mascarenhaseduardo4412@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.104.0"
uvicorn = "^0.23.2"
celery = {extras = ["rabbitmq"], version = "^5.3.4"}
pillow = "^10.1.0"
pytesseract = "^0.3.10"
openpyxl = "^3.1.2"
sqlalchemy = "^2.0.22"
psycopg2 = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
